backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.3.0
celery[redis]==5.3.6
redis==5.0.1
pydantic==2.5.3

# AI/ML
torch>=2.1.0
torchaudio>=2.1.0
openai-whisper==20231117
TTS==0.22.0
demucs==4.0.1
transformers==4.36.0
IndicTransToolkit
soundfile==0.12.1
librosa==0.10.1
numpy==1.26.3

# Speaker Detection
speechbrain==1.0.0
resemblyzer==0.1.3

# Utils
pyjwt==2.8.0